Edited field for price

+++ app/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, FileField, TextAreaField, SelectField, IntegerField
+from wtforms import StringField, SubmitField, FileField, TextAreaField, SelectField, IntegerField, DecimalField
 from wtforms.validators import InputRequired
 from flask_wtf.file import FileAllowed, FileRequired
 
@@ -8,7 +8,7 @@ class PropertyForm(FlaskForm):
     desc = TextAreaField("Description", validators=[InputRequired()])
     bedrooms = IntegerField("No. of Bedrooms", validators=[InputRequired()])
     bathrooms = IntegerField("No. of Bathrooms", validators=[InputRequired()])
-    price = IntegerField("Price", validators=[InputRequired()])
+    price = DecimalField("Price", validators=[InputRequired()])
     type = SelectField('Property Type', choices=[('house', 'House'), ('apartment', 'Apartment')], validators=[InputRequired()])
     location = StringField('Location', validators=[InputRequired()])
     photo = FileField("Photo", validators=[InputRequired(), FileRequired(), FileAllowed(['jpg','jpeg', 'jfif', 'png', 'gif'], 'Images only')])
